+++ lib/controllers/bottom_nav_controller.dart
@@ -5,24 +5,25 @@ import '../routes/page_index.dart';
 class BottomNavController extends GetxController {
   static BottomNavController get to => Get.find<BottomNavController>();
   int selectedIndex = 0;
-  List<Widget> screens = [
-    HomeScreen(),
-       if( AppController.to.basicCtrlList.isNotEmpty &&
-AppController.to.basicCtrlList[0].virtualCard.toString() == '1')
-    VirtualCardScreen(),
-      if( AppController.to.basicCtrlList.isNotEmpty &&
-AppController.to.basicCtrlList[0].virtualCard.toString() == '0')
-    TransactionScreen(),
-      if( AppController.to.basicCtrlList.isNotEmpty &&
-AppController.to.basicCtrlList[0].exchange.toString() == '1')
-    ExchangeScreen(),
-      if( AppController.to.basicCtrlList.isNotEmpty &&
-AppController.to.basicCtrlList[0].virtualCard.toString() == '0')
-    DisputeHistoryScreen(),
-    const ProfileSettingScreen(),
-  ];
+  List<Widget> get screens {
+    final basicCtrlList = AppController.to.basicCtrlList;
+    final hasVirtualCard = basicCtrlList.isNotEmpty && basicCtrlList[0].virtualCard.toString() == '1';
+    final hasExchange = basicCtrlList.isNotEmpty && basicCtrlList[0].exchange.toString() == '1';
 
-  Widget get currentScreen => screens[selectedIndex];
+    return [
+      HomeScreen(),
+      hasVirtualCard ? VirtualCardScreen() : TransactionScreen(),
+      hasExchange ? ExchangeScreen() : DisputeHistoryScreen(),
+      const ProfileSettingScreen(),
+    ];
+  }
+
+  Widget get currentScreen {
+    if (selectedIndex < 0 || selectedIndex >= screens.length) {
+      return HomeScreen(); // Fallback
+    }
+    return screens[selectedIndex];
+  }
 
   void changeScreen(int index) {
     selectedIndex = index;